refactor replace test for performance
